+++ packages/experimental/Badge/src/PresenceBadge/PresenceBadge.types.ts
@@ -1,4 +1,4 @@
-import { XmlProps } from 'react-native-svg';
+import { SvgProps } from 'react-native-svg';
 import { BadgeCoreProps, BadgeSlotProps } from '../';
 import { BadgeCoreTokens } from '../Badge.types';
 
@@ -6,14 +6,25 @@ export const presenceBadgeName = 'PresenceBadge';
 export const PresenceBadgeStatuses = ['doNotDisturb', 'busy', 'unknown', 'blocked', 'outOfOffice', 'away', 'available', 'offline'] as const;
 export type PresenceBadgeStatus = typeof PresenceBadgeStatuses[number];
 
-export interface PresenceBadgeTokens extends BadgeCoreTokens {}
+export interface PresenceBadgeTokens extends BadgeCoreTokens {
+  available?: PresenceBadgeTokens;
+  away?: PresenceBadgeTokens;
+  awayOutOfOffice?: PresenceBadgeTokens;
+  offline?: PresenceBadgeTokens;
+  outOfOffice?: PresenceBadgeTokens;
+  doNotDisturb?: PresenceBadgeTokens;
+  busy?: PresenceBadgeTokens;
+  unknown?: PresenceBadgeTokens;
+  blocked?: PresenceBadgeTokens;
+  status?: PresenceBadgeStatus;
+}
 export interface PresenceBadgeProps extends BadgeCoreProps {
   status?: PresenceBadgeStatus;
   outOfOffice?: boolean;
 }
 
 export interface PresenceBadgeSlotProps extends Omit<BadgeSlotProps, 'text' | 'icon'> {
-  svgXml: XmlProps;
+  svg: SvgProps;
 }
 
 export interface PresenceBadgeType {