+++ packages/experimental/Shadow/src/Shadow.types.ts
@@ -1,6 +1,20 @@
 import { ViewProps } from 'react-native';
 
-export type ShadowDepth = '2' | '4' | '8' | '16' | '28' | '64';
+export const shadowName = 'Shadow';
+
+export type ShadowDepth =
+  | 'shadow2'
+  | 'shadow4'
+  | 'shadow8'
+  | 'shadow16'
+  | 'shadow28'
+  | 'shadow64'
+  | 'shadow2brand'
+  | 'shadow4brand'
+  | 'shadow8brand'
+  | 'shadow16brand'
+  | 'shadow28brand'
+  | 'shadow64brand';
 
 export interface ShadowProps extends ViewProps {
   depth?: ShadowDepth;