+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Thu, 07 Jul 2022 21:24:07 GMT and should not be manually modified.
+This log was last generated on Fri, 08 Jul 2022 18:26:51 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.71.2
+
+Fri, 08 Jul 2022 18:26:51 GMT
+
+### Patches
+
+- Change shadowDepth type from '2' to 'shadow2' (78454019+lyzhan7@users.noreply.github.com)
+- Bump @fluentui-react-native/experimental-shadow to v0.2.1
+
+## 0.71.1
+
+Fri, 08 Jul 2022 17:53:30 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/avatar to v1.1.7
+- Bump @fluentui-react-native/badge to v0.0.47
 
 ## 0.71.0